fix(virto-common): merging issues
